server/requirements.txt
autopep8==1.4.4
coveralls==1.8.2
Django==2.2.4
django-extensions==2.1.9
django-nose==1.4.6
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.2
ipython==7.5.0
Pillow==6.0.0
PyJWT==1.7.1
selenium==3.141.0
tox==3.14.0
